moved _normalize_observations_and_pointings() from mapmaking/common.py to pointings.py; updated scan_map.py to use _normalize_observations_and_pointings(); updated other files accordingly
