+++ litebird_sim/mapmaking/binner.py
@@ -17,7 +17,7 @@ import healpy as hp
 from typing import Union, List, Any, Optional, Callable
 from litebird_sim.observations import Observation
 from litebird_sim.coordinates import CoordinateSystem
-from litebird_sim.pointings import get_hwp_angle
+from litebird_sim.pointings import get_hwp_angle, _normalize_observations_and_pointings
 from litebird_sim.hwp import HWP
 from litebird_sim import mpi
 from ducc0.healpix import Healpix_Base
@@ -27,7 +27,6 @@ import logging
 
 from .common import (
     _compute_pixel_indices,
-    _normalize_observations_and_pointings,
     COND_THRESHOLD,
     get_map_making_weights,
     _build_mask_detector_split,
