+++ litebird_sim/mapmaking/destriper.py
@@ -24,12 +24,11 @@ from litebird_sim.mpi import MPI_ENABLED, MPI_COMM_WORLD, MPI_COMM_GRID
 from typing import Callable, Union, List, Optional, Tuple, Any, Dict
 from litebird_sim.hwp import HWP
 from litebird_sim.observations import Observation
-from litebird_sim.pointings import get_hwp_angle
+from litebird_sim.pointings import get_hwp_angle, _normalize_observations_and_pointings
 from litebird_sim.coordinates import CoordinateSystem, coord_sys_to_healpix_string
 
 from .common import (
     _compute_pixel_indices,
-    _normalize_observations_and_pointings,
     COND_THRESHOLD,
     get_map_making_weights,
     cholesky,
